% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlotStereo.R
\name{maps2sp}
\alias{maps2sp}
\title{Transform map to sp object.}
\usage{
maps2sp(xlim, ylim, l.out = 100, proj = "+init=epsg:4326")
}
\arguments{
\item{xlim}{2-element vector with longitude limits.}

\item{ylim}{2-element vector with latitude limits.}

\item{l.out}{numeric. Length of latitudes and longitudes vector to plot.}

\item{proj}{character as passed to function \code{\link{CRS}} with the projection. Default: longlat projection.}
}
\value{
A sp object.
}
\description{
Function to slice and dice a map in longlat coordinates and convert it to an sp() object.
}
\examples{
m <-maps2sp(xlim=c(-20,40), ylim=c(20,70))
sp::spplot(m)

}
