% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overGrid.R
\name{overGrid}
\alias{overGrid}
\title{Spatial overlay for grids and Spatial* objects}
\usage{
overGrid(grid, layer, subset = FALSE)
}
\arguments{
\item{grid}{Input grid or station data.}

\item{layer}{SpatialPolygons object from which the geometries or attributes are 
queried (type \code{help(package = "sp")}.}

\item{subset}{Logical (default is FALSE). If TRUE, spatial subsetting is performed over
the imput grid using the bounding coordinates of argument 'layer'.}
}
\value{
A grid or station data.

A grid
}
\description{
Application of function \code{over} from \pkg{sp} to grids or
station data.
}
\details{
All grid locations outside layer are filled with NAs.
}
\examples{
data("EOBS_Iberia_pr")
require("visualizeR")
spatialPlot(climatology(EOBS_Iberia_pr))

data("PRUDENCEregions")
IP <- PRUDENCEregions[2,]

spatialPlot(climatology(EOBS_Iberia_pr), sp.layout = list(IP, first = FALSE))

ov <- overGrid(EOBS_Iberia_pr, layer = IP)
spatialPlot(climatology(ov), sp.layout = list(IP, first = FALSE))

ovs <- overGrid(EOBS_Iberia_pr, layer = IP, subset = TRUE)
spatialPlot(climatology(ovs), sp.layout = list(IP, first = FALSE))
}
\author{
M. Iturbide
}
