% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdf2grid.R
\name{sgdf2grid}
\alias{sgdf2grid}
\title{Convert a SpatialGridDataFrame to a C4R grid}
\usage{
sgdf2grid(sgdf, varName = NULL, level = NULL, dates = list(start =
  NULL, end = NULL), metadata = NULL)
}
\arguments{
\item{sgdf}{A SpatialGridDataFrame, as returned by function \code{gdalwarp} (package \pkg{gdalUtils}).}

\item{varName}{Character with the variable name for the resulting grid.}

\item{level}{Character with the level for the resulting grid.}

\item{dates}{Named list of the form: list(start = NULL, end = NULL), for the resulting grid.}

\item{metadata}{Named list of characters with furher metadata for the resulting grid.}
}
\value{
A C4R grid, with the metadata specified by the input arguments.
}
\description{
Convert a SpatialGridDataFrame (package \pkg{sp}) to a C4R grid
}
\seealso{
\code{\link[transformeR]{clim2sgdf}}
}
\author{
M. Iturbide
}
\keyword{internal}
