% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlotStereo.R
\name{spatialPlotStereo}
\alias{spatialPlotStereo}
\title{Spatial map with polar stereographic projection.}
\usage{
spatialPlotStereo(data, original.CRS = "+init=epsg:4326",
  new.CRS = "+init=epsg:3995", int.method = "bilinear",
  backdrop.theme = "none", set.min = floor(min(data$Data)),
  set.max = ceiling(max(data$Data)), color.theme = "RdYlBu",
  rev.colors = FALSE, ...)
}
\arguments{
\item{data}{A climatology of a grid (gridded or station dataset) or multimember grid object.}

\item{original.CRS}{character as passed to function \code{\link{CRS}} with the original projection. Default: longlat projection.}

\item{new.CRS}{character as passed to function \code{\link{CRS}} with the target projection. Default: polar stereographic projection.}

\item{int.method}{character with the interpolation method in the new projection. Default: "bilinear". See details.}

\item{backdrop.theme}{Reference geographical lines to be added to the plot. Default: "none". See details.}

\item{set.min}{Numeric value indicating an absolute minimum value (default to minimum value). All grid values below this are mapped to \code{set.min}. See details.}

\item{set.max}{Same as \code{set.min} argument, but to force a ceiling.}

\item{color.theme}{A character string indicating the color theme to use in the map. 
Valid values are those available in the \code{\link{RColorBrewer}} themes. Additionally,
the \code{"jet.colors"} palette can be used (the rainbow colors, in general not advised, though),
 for backwards compatibility. Default to the diverging, colorblind-friendly \code{"RdYlBu"} palette.
 NOTE: the \code{color.theme} argument will be overriden if the \code{col.regions} option from \code{spplot} is used.}

\item{rev.colors}{Logical. Should the chosen color theme be reversed? (default to FALSE, leaving the palette “as is”).}

\item{...}{Further arguments passed to \code{spplot}.}
}
\value{
A spatial plot using polar stereographic projection. As spplot, \code{spatialPlotStereo} returns a lattice plot of class \dQuote{trellis}.
}
\description{
Plot spatial map with polar stereographic projection.
}
\details{
The function applies the \code{\link[sp]{spplot}} method after conversion of the climatological map(s) to a
 \code{SpatialGridDataFrame}.
 
 The \code{set.min} and \code{set.max} options are useful in order to preserve adequate ranges for map representation,
  avoiding the influence of extreme values. Note that this is different than setting a range of values with an 
  interval using the \code{at} argument. The latter choice, that overrides \code{set.min} and \code{set.max},
   leaves blank grid points for outlying values.
   
\strong{Backdrop theme}
 
 Current implemented options are \code{"none"} and \code{"coastline"}, which contains
 a simplied vector theme delineating the world coastlines with the stereographic projection. Any other themes can be introduced
 by the user using the \code{sp.layout} options in \code{spplot}.

\strong{Controlling graphical parameters}
 
 Many different aspects of the map can be controlled passing the relevant arguments to 
 \code{spplot}. Fine control of graphical parameters for the trellis display can
 be also controlled using \code{\link[lattice]{trellis.par.set}}.
 
 \strong{int.method}
 
 By default bilinear interpolation is applied to get a complete grid in the target projection. Other options are "near", "cubic", "cubicspline", etc., corresponding to the argument \code{r} in \code{ gdalUtils::gdalwarp}.
}
\examples{
data(ncep_hgt500_2000)
spatialPlotStereo(climatology(ncep_hgt500_2000), backdrop.theme = "coastline")
}
\author{
A. Casanueva, J. Bedia
}
