% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdf2grid.R
\name{sgdf2grid}
\alias{sgdf2grid}
\title{Convert a SpatialGridDataFrame to a C4R grid}
\usage{
sgdf2grid(sgdf, varName = NULL, level = NULL, dates = list(start =
  NULL, end = NULL), metadata = NULL)
}
\arguments{
\item{varName}{Character}

\item{level}{Character}

\item{dates}{Named list of the form: list(start = NULL, end = NULL)}

\item{metadata}{Named list of characters}

\item{agdf}{A SpatialGridDataFrame, as returned by function...(to be completed)}
}
\value{
A grid
}
\description{
Convert a SpatialGridDataFrame (package \pkg{sp}) to a C4R grid
}
\seealso{
\code{\link[transformeR]{clim2sgdf}}
}
\author{
M. Iturbide
}
\keyword{internal}
